requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.1
cycler==0.12.1
dds-cloudapi-sdk==0.4.1
defusedxml==0.7.1
filelock==3.14.0
fonttools==4.56.0
fsspec==2025.2.0
huggingface-hub==0.29.1
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
Jinja2==3.1.5
jmespath==0.10.0
kiwisolver==1.4.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.6
model-index==0.1.11
mpmath==1.3.0
munch==4.0.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
portalocker==3.1.1
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2023.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.4.2
safetensors==0.5.3
# Editable Git install with no remote (SAM-2==1.0)
-e /home/vivora/AutoAnnotation
scipy==1.15.2
segment-anything==1.0
shapely==2.0.7
six==1.17.0
supervision==0.25.1
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
terminaltables==3.1.10
tokenizers==0.15.2
toml==0.10.2
tomli==2.2.1
torch==2.6.0
torchaudio==2.1.2+cu121
torchvision==0.21.0
tqdm==4.67.1
transformers==4.38.2
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
yapf==0.43.0
